separated common functions into separate test utils.
