+++ pycaret/tests/test_time_series_tune_random.py
@@ -7,7 +7,8 @@ import pandas as pd  # type: ignore
 
 from pycaret.datasets import get_data
 from pycaret.internal.pycaret_experiment import TimeSeriesExperiment
-from pycaret.containers.models.time_series import get_all_model_containers
+
+from .time_series_test_utils import _check_windows, _return_model_names
 
 pytestmark = pytest.mark.filterwarnings("ignore::UserWarning")
 
@@ -37,43 +38,6 @@ def load_data():
 # https://stackoverflow.com/questions/52764279/pytest-how-to-parametrize-a-test-with-a-list-that-is-returned-from-a-fixture
 # Hence, we have to create functions and create the parameterized list first
 # (must happen during collect phase) before passing it to mark.parameterize.
-
-
-def _check_windows():
-    """Check if the system is Windows."""
-    import sys
-
-    platform = sys.platform
-    is_windows = True if platform.startswith("win") else False
-
-    return is_windows
-
-
-def _return_model_names():
-    """Return all model names."""
-    globals_dict = {
-        "seed": 0,
-        "n_jobs_param": -1,
-        "gpu_param": False,
-        "X_train": pd.DataFrame(get_data("airline")),
-        "enforce_pi": False,
-    }
-    model_containers = get_all_model_containers(globals_dict)
-
-    models_to_ignore = (
-        ["prophet", "ensemble_forecaster"]
-        if _check_windows()
-        else ["ensemble_forecaster"]
-    )
-
-    model_names_ = []
-    for model_name in model_containers.keys():
-
-        if model_name not in models_to_ignore:
-            model_names_.append(model_name)
-
-    return model_names_
-
 
 _model_names = _return_model_names()
